+++ src/other/formlyUtils.js
@@ -1,0 +1,8 @@
+module.exports = {
+  setTemplates: setTemplates
+};
+
+function setTemplates(templates) {
+  console.log(templates);
+}
+