+++ CHANGELOG.md
@@ -7,9 +7,13 @@ Version numbers correspond to `bower.json` version
 - all fields will now have a the `key` value default to the index of the field if it isn't given a value. This makes it much easier to reference the model in templates (instead of `model[options.key || index]` you can now simply do `mode[options.key]`.
 - adding an angular constant called `formlyVersion` that could be useful and stuff.
 
+## Breaking changes:
+
+- in order to support `formlyUtils`, formly now exports an object instead of simply the name of the module. This is a breaking change because if you were doing `angular.module('yourModule', [require('angular-formly')]);` you will now have to change it to do `angular.module('yourModule', [require('angular-formly').ngModuleName]);`.
+
 ## Internal Changes
 
-None yet
+- Adding links to all errors and warnings. Should hopefully help people resolve issues faster.
 
 # 2.0.0